Split up archive.go into type-specific files
