+++ pkg/action/gzip.go
@@ -1,0 +1,51 @@
+package action
+
+import (
+	"compress/gzip"
+	"context"
+	"fmt"
+	"io"
+	"os"
+	"path/filepath"
+
+	"github.com/chainguard-dev/clog"
+)
+
+// extractGzip extracts .gz archives.
+func extractGzip(ctx context.Context, d string, f string) error {
+	logger := clog.FromContext(ctx).With("dir", d, "file", f)
+	logger.Debug("extracting gzip")
+
+	// Check if the file is valid
+	_, err := os.Stat(f)
+	if err != nil {
+		return fmt.Errorf("failed to stat file: %w", err)
+	}
+
+	gf, err := os.Open(f)
+	if err != nil {
+		return fmt.Errorf("failed to open file: %w", err)
+	}
+	defer gf.Close()
+
+	base := filepath.Base(f)
+	target := filepath.Join(d, base[:len(base)-len(filepath.Ext(base))])
+
+	gr, err := gzip.NewReader(gf)
+	if err != nil {
+		return fmt.Errorf("failed to create gzip reader: %w", err)
+	}
+	defer gr.Close()
+
+	ef, err := os.Create(target)
+	if err != nil {
+		return fmt.Errorf("failed to create extracted file: %w", err)
+	}
+	defer ef.Close()
+
+	if _, err := io.Copy(ef, io.LimitReader(gr, maxBytes)); err != nil {
+		return fmt.Errorf("failed to copy file: %w", err)
+	}
+
+	return nil
+}
